requirements.txt
annotated-types==0.7.0
anyio==4.8.0
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
ollama==0.4.7
pydantic==2.10.6
pydantic_core==2.27.2
PyJWT==2.10.1
requests==2.32.3
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
